refactored lib folder
