+++ src/app/[locale]/dashboard/(authenticated)/el-resources/page.tsx
@@ -1,5 +1,5 @@
 import Resources from "../../../../../components/dashboard/el_resources_page/Resources";
-import { getResouces } from "../../../../../lib/helpers";
+import { getResouces } from "../../../../../lib/helpers/regular_funcs/helpers";
 
 export default async function ElResources() {
   const data = await getResouces();
