+++ src/components/dashboard/el_resources_page/el_resource_item/actions/DecreaseCount.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { actDecreaseCount } from "../../../../../lib/actions";
+import { actDecreaseCount } from "../../../../../lib/helpers/server_act_funcs/actions";
 
 export default function DecreaseCount({ id }: { id: string }) {
   return (
