+++ src/components/ui/languagePreference/LangChanger.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import LangChangerUi from "./LangChangerUi";
-import { readCookieForClient } from "../../../lib/actions";
+import { readCookieForClient } from "../../../lib/helpers/server_act_funcs/actions";
 
 async function LangChanger() {
   const cookie = await readCookieForClient("NEXT_LOCALE");
