+++ src/components/ui/themePreference/ThemeChanger.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import ThemeChangerUi from "./ThemeChangerUi";
-import { readCookieForClient } from "../../../lib/actions";
+import { readCookieForClient } from "../../../lib/helpers/server_act_funcs/actions";
 
 async function ThemeChanger() {
   const cookie = await readCookieForClient("theme");
