+++ src/types/interfaces.d.ts
@@ -11,16 +11,6 @@ interface IUserLoginInfo {
     email: string | undefined
     password: string | undefined
 }
-interface IUserIsAuthed {
-    id: number
-    username: string
-    email: string
-    firstName: string
-    lastName: string
-    gender: string
-    image: string
-    token: string
-}
 
 // sql
 interface IUserLogin {
